+++ src/main/java/io/redstudioragnarok/valkyrie/renderer/FogRenderer.java
@@ -22,7 +22,7 @@ public class FogRenderer {
         final EntityLivingBase livingEntity = entity instanceof EntityLivingBase ? (EntityLivingBase) entity : null;
         final boolean hasBlindness = livingEntity != null && livingEntity.isPotionActive(MobEffects.BLINDNESS);
 
-        if (!ValkyrieConfig.fog.enabled && !hasBlindness) {
+        if (!ValkyrieConfig.graphics.fog.enabled && !hasBlindness) {
             GlStateManager.disableFog();
             return;
         }
@@ -66,7 +66,7 @@ public class FogRenderer {
             if (GLContext.getCapabilities().GL_NV_fog_distance)
                 GlStateManager.glFogi(34138, 34139);
         } else if (material == Material.WATER) {
-            if (!ValkyrieConfig.fog.waterFog)
+            if (!ValkyrieConfig.graphics.fog.waterFog)
                 return;
 
             GlStateManager.setFog(exp);
@@ -76,13 +76,13 @@ public class FogRenderer {
             else
                 GlStateManager.setFogDensity(0.1F);
         } else if (material == Material.LAVA) {
-            if (!ValkyrieConfig.fog.lavaFog)
+            if (!ValkyrieConfig.graphics.fog.lavaFog)
                 return;
 
             GlStateManager.setFog(exp);
             GlStateManager.setFogDensity(2);
         } else {
-            if (!ValkyrieConfig.fog.distanceFog) {
+            if (!ValkyrieConfig.graphics.fog.distanceFog) {
                 GlStateManager.disableFog();
                 return;
             }